requirements.txt
numpy
omegaconf
msgspec
rouge-score
sacrebleu
sentencepiece
tiktoken
blobfile
wandb
viztracer
lm-eval
scipy
pynvml
orjson
einops
pillow
pyahocorasick
iopath
torchdata
torchcodec
timm
decord
opencv-python
pycocoevalcap
scikit-learn
scipy
sentencepiece
tokenizers
webdataset
fsspec
datatrove
ftfy
transformers>=4.48.0
